+++ openapi-cli/src/test/java/io/ballerina/openapi/generators/client/FunctionBodyNodeTests.java
@@ -66,7 +66,7 @@ public class FunctionBodyNodeTests {
         TypeHandler.createInstance(openapi, false);
         FunctionBodyGeneratorImp functionBodyGeneratorImp = new FunctionBodyGeneratorImp(path, operation, openapi,
                 new AuthConfigGeneratorImp(false, false),
-                new BallerinaUtilGenerator(), new ArrayList<>());
+                new BallerinaUtilGenerator(), new ArrayList<>(), false, false, false);
         Optional<FunctionBodyNode> bodyNode = functionBodyGeneratorImp.getFunctionBodyNode();
         content = content.trim().replaceAll("\n", "").replaceAll("\\s+", "");
         String bodyNodeContent = bodyNode.get().toString().trim().replaceAll("\n", "")
