+++ openapi-cli/src/test/java/io/ballerina/openapi/generators/client/FunctionSignatureNodeTests.java
@@ -61,7 +61,7 @@ public class FunctionSignatureNodeTests {
         Operation operation = openAPI.getPaths().get("/products/{country}").getGet();
         TypeHandler.createInstance(openAPI, false);
         RemoteFunctionSignatureGenerator functionSignatureGenerator = new RemoteFunctionSignatureGenerator(operation,
-                openAPI, "get");
+                openAPI, "get", "products/{country}");
         Optional<FunctionSignatureNode> signature = functionSignatureGenerator.generateFunctionSignature();
         SeparatedNodeList<ParameterNode> parameters = signature.get().parameters();
         Assert.assertFalse(parameters.isEmpty());
@@ -92,7 +92,7 @@ public class FunctionSignatureNodeTests {
         Operation operation = openAPI.getPaths().get("/pets").getPost();
         TypeHandler.createInstance(openAPI, false);
         RemoteFunctionSignatureGenerator functionSignatureGenerator = new RemoteFunctionSignatureGenerator(operation,
-                 openAPI, "post");
+                 openAPI, "post", "/pets");
         FunctionSignatureNode signature = functionSignatureGenerator.generateFunctionSignature().get();
         SeparatedNodeList<ParameterNode> parameters = signature.parameters();
         Assert.assertFalse(parameters.isEmpty());
@@ -111,7 +111,7 @@ public class FunctionSignatureNodeTests {
         Operation operation = openAPI.getPaths().get("/pets").getPost();
         TypeHandler.createInstance(openAPI, false);
         RemoteFunctionSignatureGenerator functionSignatureGenerator = new RemoteFunctionSignatureGenerator(operation,
-                openAPI, "post");
+                openAPI, "post", "/pets");
         FunctionSignatureNode signature = functionSignatureGenerator.generateFunctionSignature().get();
         SeparatedNodeList<ParameterNode> parameters = signature.parameters();
         Assert.assertFalse(parameters.isEmpty());
@@ -131,7 +131,7 @@ public class FunctionSignatureNodeTests {
         Operation operation = openAPI.getPaths().get("/pets").getPost();
         TypeHandler.createInstance(openAPI, false);
         RemoteFunctionSignatureGenerator functionSignatureGenerator = new RemoteFunctionSignatureGenerator(operation,
-                openAPI, "post");
+                openAPI, "post", "/pets");
         FunctionSignatureNode signature = functionSignatureGenerator.generateFunctionSignature().get();
         SeparatedNodeList<ParameterNode> parameters = signature.parameters();
         Assert.assertFalse(parameters.isEmpty());
@@ -157,7 +157,7 @@ public class FunctionSignatureNodeTests {
         OpenAPI openAPI = getOpenAPI(RESDIR.resolve("swagger/response_nested_array.yaml"));
         TypeHandler.createInstance(openAPI, false);
         RemoteFunctionSignatureGenerator functionSignatureGenerator = new RemoteFunctionSignatureGenerator(
-                openAPI.getPaths().get("/timestags").getGet(), openAPI, "get");
+                openAPI.getPaths().get("/timestags").getGet(), openAPI, "get", "/timestags");
         FunctionSignatureNode signature = functionSignatureGenerator.generateFunctionSignature().get();
         ReturnTypeDescriptorNode returnTypeNode = signature.returnTypeDesc().orElseThrow();
         Assert.assertEquals(returnTypeNode.type().toString(), "string[][]|error");
@@ -168,7 +168,7 @@ public class FunctionSignatureNodeTests {
         OpenAPI openAPI = getOpenAPI(RESDIR.resolve("swagger/response_string_array.yaml"));
         TypeHandler.createInstance(openAPI, false);
         RemoteFunctionSignatureGenerator functionSignatureGenerator = new RemoteFunctionSignatureGenerator(
-                openAPI.getPaths().get("/timestags").getGet(), openAPI, "get");
+                openAPI.getPaths().get("/timestags").getGet(), openAPI, "get", "/timestags");
         FunctionSignatureNode signature = functionSignatureGenerator.generateFunctionSignature().get();
         ReturnTypeDescriptorNode returnTypeNode = signature.returnTypeDesc().orElseThrow();
         Assert.assertEquals(returnTypeNode.type().toString(), "string[]|error");
@@ -179,7 +179,7 @@ public class FunctionSignatureNodeTests {
         OpenAPI openAPI = getOpenAPI(RESDIR.resolve("swagger/request_body_with_ref.yaml"));
         TypeHandler.createInstance(openAPI, false);
         RemoteFunctionSignatureGenerator functionSignatureGenerator = new RemoteFunctionSignatureGenerator(
-                openAPI.getPaths().get("/pets").getPost(), openAPI, "post");
+                openAPI.getPaths().get("/pets").getPost(), openAPI, "post", "/pets");
         FunctionSignatureNode signature = functionSignatureGenerator.generateFunctionSignature().get();
         SeparatedNodeList<ParameterNode> parameters = signature.parameters();
         Assert.assertFalse(parameters.isEmpty());
@@ -193,7 +193,7 @@ public class FunctionSignatureNodeTests {
         OpenAPI openAPI = getOpenAPI(RESDIR.resolve("swagger/pdf_payload.yaml"));
         TypeHandler.createInstance(openAPI, false);
         RemoteFunctionSignatureGenerator functionSignatureGenerator = new RemoteFunctionSignatureGenerator(
-                openAPI.getPaths().get("/pets").getPost(), openAPI, "post");
+                openAPI.getPaths().get("/pets").getPost(), openAPI, "post", "/pets");
         FunctionSignatureNode signature = functionSignatureGenerator.generateFunctionSignature().get();
         SeparatedNodeList<ParameterNode> parameters = signature.parameters();
         Assert.assertFalse(parameters.isEmpty());
@@ -210,7 +210,7 @@ public class FunctionSignatureNodeTests {
         OpenAPI openAPI = getOpenAPI(RESDIR.resolve("swagger/invalid_array_query_params.yaml"));
         TypeHandler.createInstance(openAPI, false);
         RemoteFunctionSignatureGenerator functionSignatureGenerator = new RemoteFunctionSignatureGenerator(
-                openAPI.getPaths().get("/pets").getPost(), openAPI, "post");
+                openAPI.getPaths().get("/pets").getPost(), openAPI, "post", "/pets");
         FunctionSignatureNode signature = functionSignatureGenerator.generateFunctionSignature().get();
 
     }
@@ -223,7 +223,7 @@ public class FunctionSignatureNodeTests {
         OpenAPI openAPI = getOpenAPI(RESDIR.resolve("swagger/invalid_array_query_params.yaml"));
         TypeHandler.createInstance(openAPI, false);
         RemoteFunctionSignatureGenerator functionSignatureGenerator = new RemoteFunctionSignatureGenerator(
-                openAPI.getPaths().get("/dogs").getGet(), openAPI, "get");
+                openAPI.getPaths().get("/dogs").getGet(), openAPI, "get", "/dogs");
         FunctionSignatureNode signature = functionSignatureGenerator.generateFunctionSignature().get();
 
     }
@@ -233,10 +233,10 @@ public class FunctionSignatureNodeTests {
         OpenAPI openAPI = getOpenAPI(RESDIR.resolve("swagger/integer_request_payload.yaml"));
         TypeHandler.createInstance(openAPI, false);
         RemoteFunctionSignatureGenerator signature = new RemoteFunctionSignatureGenerator(openAPI.getPaths().
-                get("/pets").getPost(), openAPI, "post");
+                get("/pets").getPost(), openAPI, "post", "/pets");
         FunctionSignatureNode petSignature = signature.generateFunctionSignature().get();
         RemoteFunctionSignatureGenerator owSignature = new RemoteFunctionSignatureGenerator(openAPI.getPaths().
-                get("/owners").getPost(), openAPI, "post");
+                get("/owners").getPost(), openAPI, "post", "/owners");
         FunctionSignatureNode ownerSignature = owSignature.generateFunctionSignature().get();
         SeparatedNodeList<ParameterNode> parameters = petSignature.parameters();
         Assert.assertFalse(parameters.isEmpty());
