+++ openapi-core/src/main/java/io/ballerina/openapi/core/generators/client/FunctionBodyGeneratorImp.java
@@ -34,7 +34,6 @@ import io.ballerina.compiler.syntax.tree.Node;
 import io.ballerina.compiler.syntax.tree.NodeList;
 import io.ballerina.compiler.syntax.tree.ReturnStatementNode;
 import io.ballerina.compiler.syntax.tree.SimpleNameReferenceNode;
-import io.ballerina.compiler.syntax.tree.SpecificFieldNode;
 import io.ballerina.compiler.syntax.tree.StatementNode;
 import io.ballerina.compiler.syntax.tree.TemplateExpressionNode;
 import io.ballerina.compiler.syntax.tree.Token;
@@ -83,14 +82,11 @@ import static io.ballerina.compiler.syntax.tree.NodeFactory.createIfElseStatemen
 import static io.ballerina.compiler.syntax.tree.NodeFactory.createMappingConstructorExpressionNode;
 import static io.ballerina.compiler.syntax.tree.NodeFactory.createReturnStatementNode;
 import static io.ballerina.compiler.syntax.tree.NodeFactory.createSimpleNameReferenceNode;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createSpecificFieldNode;
 import static io.ballerina.compiler.syntax.tree.NodeFactory.createTemplateExpressionNode;
 import static io.ballerina.compiler.syntax.tree.NodeFactory.createTypedBindingPatternNode;
 import static io.ballerina.compiler.syntax.tree.NodeFactory.createVariableDeclarationNode;
 import static io.ballerina.compiler.syntax.tree.SyntaxKind.BACKTICK_TOKEN;
 import static io.ballerina.compiler.syntax.tree.SyntaxKind.CLOSE_BRACE_TOKEN;
-import static io.ballerina.compiler.syntax.tree.SyntaxKind.COLON_TOKEN;
-import static io.ballerina.compiler.syntax.tree.SyntaxKind.COMMA_TOKEN;
 import static io.ballerina.compiler.syntax.tree.SyntaxKind.DOT_TOKEN;
 import static io.ballerina.compiler.syntax.tree.SyntaxKind.EQUAL_TOKEN;
 import static io.ballerina.compiler.syntax.tree.SyntaxKind.IF_KEYWORD;
@@ -105,7 +101,7 @@ import static io.ballerina.openapi.core.generators.common.GeneratorConstants.DEL
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.ENCODING;
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.EXECUTE;
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.HEAD;
-import static io.ballerina.openapi.core.generators.common.GeneratorConstants.HEADER;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.HEADERS;
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.HEADER_VALUES;
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.HTTP_HEADERS;
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.HTTP_REQUEST;
@@ -113,7 +109,7 @@ import static io.ballerina.openapi.core.generators.common.GeneratorConstants.NEW
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.PATCH;
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.POST;
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.PUT;
-import static io.ballerina.openapi.core.generators.common.GeneratorConstants.QUERY;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.QUERIES;
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.QUERY_PARAM;
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.REQUEST;
 import static io.ballerina.openapi.core.generators.common.GeneratorConstants.RESOURCE_PATH;
@@ -133,29 +129,35 @@ import static io.ballerina.openapi.core.generators.common.GeneratorUtils.isCompo
 public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
 
     private List<ImportDeclarationNode> imports = new ArrayList<>();
-    private boolean isHeader;
     private final String path;
     private final Map.Entry<PathItem.HttpMethod, Operation> operation;
     private final OpenAPI openAPI;
     private final BallerinaUtilGenerator ballerinaUtilGenerator;
     private final AuthConfigGeneratorImp ballerinaAuthConfigGeneratorImp;
 
+    private final boolean hasHeaders;
+    private final boolean hasQueries;
+    private final boolean hasDefaultHeaders;
+
     public List<ImportDeclarationNode> getImports() {
         return imports;
     }
 
 
-    public FunctionBodyGeneratorImp(String path, Map.Entry<PathItem.HttpMethod, Operation> operation, OpenAPI openAPI,
-                                    AuthConfigGeneratorImp ballerinaAuthConfigGeneratorImp,
+    public FunctionBodyGeneratorImp(String path, Map.Entry<PathItem.HttpMethod, Operation> operation,
+                                    OpenAPI openAPI, AuthConfigGeneratorImp ballerinaAuthConfigGeneratorImp,
                                     BallerinaUtilGenerator ballerinaUtilGenerator,
-                                    List<ImportDeclarationNode> imports) {
+                                    List<ImportDeclarationNode> imports, boolean hasHeaders,
+                                    boolean hasDefaultHeaders, boolean hasQueries) {
         this.path = path;
         this.operation = operation;
-        this.isHeader = false;
         this.openAPI = openAPI;
         this.ballerinaUtilGenerator = ballerinaUtilGenerator;
         this.ballerinaAuthConfigGeneratorImp = ballerinaAuthConfigGeneratorImp;
         this.imports = imports;
+        this.hasHeaders = hasHeaders;
+        this.hasQueries = hasQueries;
+        this.hasDefaultHeaders = hasDefaultHeaders;
     }
 
     /**
@@ -168,31 +170,29 @@ public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
     public Optional<FunctionBodyNode> getFunctionBodyNode() {
 
         NodeList<AnnotationNode> annotationNodes = createEmptyNodeList();
-        isHeader = false;
         // Create statements
         List<StatementNode> statementsList = new ArrayList<>();
         //string path - common for every remote functions
         VariableDeclarationNode pathInt = getPathStatement(path, annotationNodes);
         statementsList.add(pathInt);
         try {
+            //Handle query parameter map
+            handleParameterSchemaInOperation(operation, statementsList);
 
-            //Handel query parameter map
-        handleParameterSchemaInOperation(operation, statementsList);
+            String method = operation.getKey().name().trim().toLowerCase(Locale.ENGLISH);
+            // Statement Generator for requestBody
+            if (operation.getValue().getRequestBody() != null) {
+                RequestBody requestBody = operation.getValue().getRequestBody();
+                handleRequestBodyInOperation(statementsList, method, requestBody);
+            } else {
+                createCommonFunctionBodyStatements(statementsList, method);
+            }
 
-        String method = operation.getKey().name().trim().toLowerCase(Locale.ENGLISH);
-        // Statement Generator for requestBody
-        if (operation.getValue().getRequestBody() != null) {
-            RequestBody requestBody = operation.getValue().getRequestBody();
-            handleRequestBodyInOperation(statementsList, method, requestBody);
-        } else {
-            createCommonFunctionBodyStatements(statementsList, method);
-        }
-
-        //Create statements
-        NodeList<StatementNode> statements = createNodeList(statementsList);
-        return Optional.of(createFunctionBodyBlockNode(createToken(OPEN_BRACE_TOKEN), null,
-                statements,
-                createToken(CLOSE_BRACE_TOKEN), null));
+            //Create statements
+            NodeList<StatementNode> statements = createNodeList(statementsList);
+            return Optional.of(createFunctionBodyBlockNode(createToken(OPEN_BRACE_TOKEN), null,
+                    statements,
+                    createToken(CLOSE_BRACE_TOKEN), null));
         } catch (BallerinaOpenApiException e) {
             //todo diagnostic message
             return Optional.empty();
@@ -226,20 +226,6 @@ public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
         List<Parameter> queryParameters = new ArrayList<>();
         List<Parameter> headerParameters = new ArrayList<>();
 
-        if (operation.getValue().getParameters() != null) {
-            List<Parameter> parameters = operation.getValue().getParameters();
-            for (Parameter parameter : parameters) {
-                if (parameter.get$ref() != null) {
-                    String[] splits = parameter.get$ref().split("/");
-                    parameter = openAPI.getComponents().getParameters().get(splits[splits.length - 1]);
-                }
-                if (parameter.getIn().trim().equals(QUERY)) {
-                    queryParameters.add(parameter);
-                } else if (parameter.getIn().trim().equals(HEADER)) {
-                    headerParameters.add(parameter);
-                }
-            }
-        }
         handleQueryParamsAndHeaders(queryParameters, headerParameters, statementsList, queryApiKeyNameList,
                 headerApiKeyNameList);
     }
@@ -255,20 +241,20 @@ public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
                 ballerinaAuthConfigGeneratorImp.isApiKey();
         if (combinationOfApiKeyAndHTTPOAuth) {
             addUpdatedPathAndHeaders(statementsList, queryApiKeyNameList, queryParameters,
-                    headerApiKeyNameList, headerParameters);
+                    headerApiKeyNameList);
         } else {
-            if (!queryParameters.isEmpty() || !queryApiKeyNameList.isEmpty()) {
+            if (hasQueries || !queryApiKeyNameList.isEmpty()) {
                 ballerinaUtilGenerator.setQueryParamsFound(true);
-                statementsList.add(getMapForParameters(queryParameters, "map<anydata>",
-                        QUERY_PARAM, queryApiKeyNameList));
                 getUpdatedPathHandlingQueryParamEncoding(statementsList, queryParameters);
             }
-            if (!headerParameters.isEmpty() || !headerApiKeyNameList.isEmpty()) {
-                statementsList.add(getMapForParameters(headerParameters, "map<anydata>",
-                        HEADER_VALUES, headerApiKeyNameList));
-                statementsList.add(GeneratorUtils.getSimpleExpressionStatementNode(
-                        "map<string|string[]> " + HTTP_HEADERS + " = getMapForHeaders(headerValues)"));
-                isHeader = true;
+            if (hasHeaders || !headerApiKeyNameList.isEmpty()) {
+                if (hasDefaultHeaders) {
+                    statementsList.add(GeneratorUtils.getSimpleExpressionStatementNode(
+                            "map<string|string[]> " + HTTP_HEADERS + " = " + HEADERS));
+                } else {
+                    statementsList.add(GeneratorUtils.getSimpleExpressionStatementNode(
+                            "map<string|string[]> " + HTTP_HEADERS + " = getMapForHeaders(" + HEADERS + ")"));
+                }
                 ballerinaUtilGenerator.setHeadersFound(true);
             }
         }
@@ -279,54 +265,53 @@ public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
      * authentication.
      */
     private void addUpdatedPathAndHeaders(List<StatementNode> statementsList, List<String> queryApiKeyNameList,
-                                          List<Parameter> queryParameters, List<String> headerApiKeyNameList,
-                                          List<Parameter> headerParameters) throws BallerinaOpenApiException {
+                                          List<Parameter> queryParameters, List<String> headerApiKeyNameList)
+            throws BallerinaOpenApiException {
 
         List<StatementNode> ifBodyStatementsList = new ArrayList<>();
 
-        if (!headerParameters.isEmpty() || !headerApiKeyNameList.isEmpty()) {
-            if (!headerParameters.isEmpty()) {
-                statementsList.add(getMapForParameters(headerParameters, "map<any>",
-                        HEADER_VALUES, new ArrayList<>()));
-            } else {
-                ExpressionStatementNode headerMapCreation = GeneratorUtils.getSimpleExpressionStatementNode(
-                        "map<any> " + HEADER_VALUES + " = {}");
-                statementsList.add(headerMapCreation);
+        if (hasHeaders || !headerApiKeyNameList.isEmpty()) {
+            String defaultValue = "{}";
+            if (hasHeaders) {
+                defaultValue = "{..." + HEADERS + "}";
             }
+
+            ExpressionStatementNode headerMapCreation = GeneratorUtils.getSimpleExpressionStatementNode(
+                    "map<anydata> " + HEADER_VALUES + " = " + defaultValue);
+            statementsList.add(headerMapCreation);
 
             if (!headerApiKeyNameList.isEmpty()) {
                 // update headerValues Map within the if block
                 // `headerValues["api-key"] = self.apiKeyConfig?.apiKey;`
                 addApiKeysToMap(HEADER_VALUES, headerApiKeyNameList, ifBodyStatementsList);
             }
-            isHeader = true;
             ballerinaUtilGenerator.setHeadersFound(true);
         }
 
-        if (!queryParameters.isEmpty() || !queryApiKeyNameList.isEmpty()) {
-            ballerinaUtilGenerator.setQueryParamsFound(true);
-            if (!queryParameters.isEmpty()) {
-                statementsList.add(getMapForParameters(queryParameters, "map<anydata>",
-                        QUERY_PARAM, new ArrayList<>()));
-            } else {
-                ExpressionStatementNode queryParamMapCreation = GeneratorUtils.getSimpleExpressionStatementNode(
-                        "map<anydata> " + QUERY_PARAM + " = {}");
-                statementsList.add(queryParamMapCreation);
+        if (hasQueries || !queryApiKeyNameList.isEmpty()) {
+            String defaultValue = "{}";
+            if (hasQueries) {
+                defaultValue = "{..." + QUERIES + "}";
             }
+
+            ExpressionStatementNode queryParamMapCreation = GeneratorUtils.getSimpleExpressionStatementNode(
+                    "map<anydata> " + QUERY_PARAM + " = " + defaultValue);
+            statementsList.add(queryParamMapCreation);
 
             if (!queryApiKeyNameList.isEmpty()) {
                 // update queryParam Map within the if block
                 // `queryParam["api-key"] = self.apiKeyConfig?.apiKey;`
                 addApiKeysToMap(QUERY_PARAM, queryApiKeyNameList, ifBodyStatementsList);
             }
+            ballerinaUtilGenerator.setQueryParamsFound(true);
         }
 
         generateIfBlockToAddApiKeysToMaps(statementsList, ifBodyStatementsList);
 
-        if (!queryParameters.isEmpty() || !queryApiKeyNameList.isEmpty()) {
+        if (hasQueries || !queryApiKeyNameList.isEmpty()) {
             getUpdatedPathHandlingQueryParamEncoding(statementsList, queryParameters);
         }
-        if (!headerParameters.isEmpty() || !headerApiKeyNameList.isEmpty()) {
+        if (hasHeaders || !headerApiKeyNameList.isEmpty()) {
             statementsList.add(GeneratorUtils.getSimpleExpressionStatementNode(
                     "map<string|string[]> " + HTTP_HEADERS + " = getMapForHeaders(headerValues)"));
         }
@@ -368,12 +353,12 @@ public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
         if (queryParamEncodingMap != null) {
             statementsList.add(queryParamEncodingMap);
             ExpressionStatementNode updatedPath = GeneratorUtils.getSimpleExpressionStatementNode(
-                    RESOURCE_PATH + " = " + RESOURCE_PATH + " + check getPathForQueryParam(queryParam, " +
+                    RESOURCE_PATH + " = " + RESOURCE_PATH + " + check getPathForQueryParam(" + QUERIES + ", " +
                             "queryParamEncoding)");
             statementsList.add(updatedPath);
         } else {
             ExpressionStatementNode updatedPath = GeneratorUtils.getSimpleExpressionStatementNode(
-                    RESOURCE_PATH + " = " + RESOURCE_PATH + " + check getPathForQueryParam(queryParam)");
+                    RESOURCE_PATH + " = " + RESOURCE_PATH + " + check getPathForQueryParam(" + QUERIES + ")");
             statementsList.add(updatedPath);
         }
     }
@@ -493,7 +478,7 @@ public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
             Iterator<Map.Entry<String, MediaType>> iterator = entries.iterator();
             //Currently align with first content of the requestBody
             while (iterator.hasNext()) {
-                createRequestBodyStatements(isHeader, statementsList, method, iterator);
+                createRequestBodyStatements(statementsList, method, iterator);
                 break;
             }
         } else if (requestBody.get$ref() != null) {
@@ -504,7 +489,7 @@ public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
             Iterator<Map.Entry<String, MediaType>> iterator = entries.iterator();
             //Currently align with first content of the requestBody
             while (iterator.hasNext()) {
-                createRequestBodyStatements(isHeader, statementsList, method, iterator);
+                createRequestBodyStatements(statementsList, method, iterator);
                 break;
             }
         }
@@ -520,7 +505,7 @@ public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
         // This condition for several methods.
         boolean isEntityBodyMethods = method.equals(POST) || method.equals(PUT) || method.equals(PATCH)
                 || method.equals(EXECUTE);
-        if (isHeader) {
+        if (hasHeaders) {
             if (isEntityBodyMethods) {
                 ExpressionStatementNode requestStatementNode = GeneratorUtils.getSimpleExpressionStatementNode(
                         "http:Request request = new");
@@ -635,12 +620,11 @@ public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
      *    self.clientEp->put(path, request);
      * </pre>
      *
-     * @param isHeader       - Boolean value for header availability.
      * @param statementsList - StatementNode list in body node
      * @param method         - Operation method name.
      * @param iterator       - RequestBody media type
      */
-    private void createRequestBodyStatements(boolean isHeader, List<StatementNode> statementsList, String method,
+    private void createRequestBodyStatements(List<StatementNode> statementsList, String method,
                                              Iterator<Map.Entry<String, MediaType>> iterator)
             throws BallerinaOpenApiException {
 
@@ -662,7 +646,7 @@ public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
         }
         // POST, PUT, PATCH, DELETE, EXECUTE
         String requestStatement = getClientCallWithRequest().formatted(method, RESOURCE_PATH);
-        if (isHeader) {
+        if (hasHeaders) {
             if (method.equals(POST) || method.equals(PUT) || method.equals(PATCH) || method.equals(DELETE)
                     || method.equals(EXECUTE)) {
                 requestStatement = getClientCallWithRequestAndHeaders().formatted(method, RESOURCE_PATH,
@@ -701,64 +685,6 @@ public class FunctionBodyGeneratorImp implements FunctionBodyGenerator {
         MimeFactory factory = new MimeFactory();
         MimeType mimeType = factory.getMimeType(mediaTypeEntry, ballerinaUtilGenerator, imports);
         mimeType.setPayload(statementsList, mediaTypeEntry);
-    }
-
-    /**
-     * Generate map variable for query parameters and headers.
-     */
-    private VariableDeclarationNode getMapForParameters(List<Parameter> parameters, String mapDataType,
-                                                        String mapName, List<String> apiKeyNames) {
-        List<Node> filedOfMap = new ArrayList<>();
-        BuiltinSimpleNameReferenceNode mapType = createBuiltinSimpleNameReferenceNode(null,
-                createIdentifierToken(mapDataType));
-        CaptureBindingPatternNode bindingPattern = createCaptureBindingPatternNode(
-                createIdentifierToken(mapName));
-        TypedBindingPatternNode bindingPatternNode = createTypedBindingPatternNode(mapType, bindingPattern);
-
-        for (Parameter parameter : parameters) {
-            // Initializer
-            IdentifierToken fieldName = createIdentifierToken('"' + (parameter.getName().trim()) + '"');
-            Token colon = createToken(COLON_TOKEN);
-            SimpleNameReferenceNode valueExpr = createSimpleNameReferenceNode(
-                    createIdentifierToken(escapeIdentifier(parameter.getName().trim())));
-            SpecificFieldNode specificFieldNode = createSpecificFieldNode(null,
-                    fieldName, colon, valueExpr);
-            filedOfMap.add(specificFieldNode);
-            filedOfMap.add(createToken(COMMA_TOKEN));
-        }
-
-        if (!apiKeyNames.isEmpty()) {
-            for (String apiKey : apiKeyNames) {
-                IdentifierToken fieldName = createIdentifierToken('"' + apiKey.trim() + '"');
-                Token colon = createToken(COLON_TOKEN);
-                IdentifierToken apiKeyConfigIdentifierToken = createIdentifierToken(API_KEY_CONFIG_PARAM);
-                if (ballerinaAuthConfigGeneratorImp.isHttpOROAuth() && ballerinaAuthConfigGeneratorImp.isApiKey()) {
-                    apiKeyConfigIdentifierToken = createIdentifierToken(API_KEY_CONFIG_PARAM +
-                            QUESTION_MARK_TOKEN.stringValue());
-                }
-                SimpleNameReferenceNode apiKeyConfigParamNode = createSimpleNameReferenceNode(
-                        apiKeyConfigIdentifierToken);
-                FieldAccessExpressionNode fieldExpr = createFieldAccessExpressionNode(
-                        createSimpleNameReferenceNode(createIdentifierToken(SELF)), createToken(DOT_TOKEN),
-                        apiKeyConfigParamNode);
-                SimpleNameReferenceNode valueExpr = createSimpleNameReferenceNode(createIdentifierToken(
-                        getValidName(apiKey, false)));
-                SpecificFieldNode specificFieldNode;
-                ExpressionNode apiKeyExpr = createFieldAccessExpressionNode(
-                        fieldExpr, createToken(DOT_TOKEN), valueExpr);
-                specificFieldNode = createSpecificFieldNode(null, fieldName, colon, apiKeyExpr);
-                filedOfMap.add(specificFieldNode);
-                filedOfMap.add(createToken(COMMA_TOKEN));
-            }
-        }
-
-        filedOfMap.remove(filedOfMap.size() - 1);
-        MappingConstructorExpressionNode initialize = createMappingConstructorExpressionNode(
-                createToken(OPEN_BRACE_TOKEN), createSeparatedNodeList(filedOfMap),
-                createToken(CLOSE_BRACE_TOKEN));
-        return createVariableDeclarationNode(createEmptyNodeList(),
-                null, bindingPatternNode, createToken(EQUAL_TOKEN), initialize,
-                createToken(SEMICOLON_TOKEN));
     }
 
     @Override
