+++ openapi-core/src/main/java/io/ballerina/openapi/core/generators/client/ResourceFunctionSignatureGenerator.java
@@ -18,7 +18,6 @@
 
 package io.ballerina.openapi.core.generators.client;
 
-import io.ballerina.compiler.syntax.tree.DefaultableParameterNode;
 import io.ballerina.compiler.syntax.tree.FunctionSignatureNode;
 import io.ballerina.compiler.syntax.tree.Node;
 import io.ballerina.compiler.syntax.tree.NodeFactory;
@@ -29,9 +28,8 @@ import io.ballerina.compiler.syntax.tree.SeparatedNodeList;
 import io.ballerina.compiler.syntax.tree.Token;
 import io.ballerina.openapi.core.generators.client.diagnostic.ClientDiagnostic;
 import io.ballerina.openapi.core.generators.client.diagnostic.ClientDiagnosticImp;
-import io.ballerina.openapi.core.generators.client.diagnostic.DiagnosticMessages;
-import io.ballerina.openapi.core.generators.client.parameter.HeaderParameterGenerator;
-import io.ballerina.openapi.core.generators.client.parameter.QueryParameterGenerator;
+import io.ballerina.openapi.core.generators.client.parameter.HeadersParameterGenerator;
+import io.ballerina.openapi.core.generators.client.parameter.QueriesParameterGenerator;
 import io.ballerina.openapi.core.generators.client.parameter.RequestBodyGenerator;
 import io.ballerina.openapi.core.generators.common.exception.BallerinaOpenApiException;
 import io.swagger.v3.oas.models.OpenAPI;
@@ -55,10 +53,19 @@ public class ResourceFunctionSignatureGenerator implements FunctionSignatureGene
     protected final OpenAPI openAPI;
     protected final List<ClientDiagnostic> diagnostics = new ArrayList<>();
     protected FunctionReturnTypeGeneratorImp functionReturnTypeGenerator;
+    private final String httpMethod;
+    private final String path;
 
-    public ResourceFunctionSignatureGenerator(Operation operation, OpenAPI openAPI, String httpMethod) {
+    private boolean hasDefaultHeader = false;
+    private boolean hasHeadersParam = false;
+    private boolean hasQueriesParam = false;
+
+    public ResourceFunctionSignatureGenerator(Operation operation, OpenAPI openAPI, String httpMethod,
+                                              String path) {
         this.operation = operation;
         this.openAPI = openAPI;
+        this.httpMethod = httpMethod;
+        this.path = path;
         this.functionReturnTypeGenerator = new FunctionReturnTypeGeneratorImp(operation, openAPI, httpMethod);
     }
 
@@ -99,93 +106,106 @@ public class ResourceFunctionSignatureGenerator implements FunctionSignatureGene
     }
 
     protected ParametersInfo getParametersInfo(List<Parameter> parameters) {
-        // 1. parameters - path , query, requestBody, headers
-        List<String> paramName = new ArrayList<>();
         List<Node> parameterList = new ArrayList<>();
         List<Node> defaultable = new ArrayList<>();
         Token comma = createToken(COMMA_TOKEN);
-        if (parameters != null) {
-            for (Parameter parameter : parameters) {
-                if (parameter.get$ref() != null) {
-                    String paramType = null;
-                    try {
-                        paramType = extractReferenceType(parameter.get$ref());
-                    } catch (BallerinaOpenApiException e) {
-                        DiagnosticMessages diagnostic = OAS_CLIENT_100;
-                        ClientDiagnosticImp clientDiagnostic = new ClientDiagnosticImp(diagnostic,
-                                parameter.get$ref());
-                        diagnostics.add(clientDiagnostic);
-                    }
-                    parameter = openAPI.getComponents().getParameters().get(paramType);
-                }
 
-                String in = parameter.getIn();
+        List<Parameter> headerParameters = new ArrayList<>();
+        List<Parameter> queryParameters = new ArrayList<>();
 
-                switch (in) {
-                    case "query":
-                        QueryParameterGenerator queryParameterGenerator = new QueryParameterGenerator(parameter,
-                                openAPI);
-                        Optional<ParameterNode> queryParam = queryParameterGenerator.generateParameterNode();
-                        if (queryParam.isEmpty()) {
-                            diagnostics.addAll(queryParameterGenerator.getDiagnostics());
-                            return null;
-                        }
-                        if (queryParam.get() instanceof RequiredParameterNode requiredParameterNode) {
-                            parameterList.add(requiredParameterNode);
-                            parameterList.add(comma);
-                        } else {
-                            defaultable.add(queryParam.get());
-                            defaultable.add(comma);
-                        }
-                        break;
-                    case "header":
-                        HeaderParameterGenerator headerParameterGenerator = new HeaderParameterGenerator(parameter,
-                                openAPI);
-                        Optional<ParameterNode> headerParam = headerParameterGenerator.generateParameterNode();
-                        if (headerParam.isEmpty()) {
-                            diagnostics.addAll(headerParameterGenerator.getDiagnostics());
-                            return null;
-                        }
-                        if (headerParam.get() instanceof RequiredParameterNode headerNode) {
-                            parameterList.add(headerNode);
-                            parameterList.add(comma);
-                        } else {
-                            defaultable.add(headerParam.get());
-                            defaultable.add(comma);
-                        }
-                        break;
-                    default:
-                        break;
-                }
-            }
-        }
-        // 2. requestBody
+        // 1. requestBody
         if (operation.getRequestBody() != null) {
-            RequestBodyGenerator requestBodyGenerator = new RequestBodyGenerator(operation.getRequestBody(), openAPI);
+            RequestBodyGenerator requestBodyGenerator = new RequestBodyGenerator(operation.getRequestBody(),
+                    openAPI);
             Optional<ParameterNode> requestBody = requestBodyGenerator.generateParameterNode();
             if (requestBody.isEmpty()) {
                 diagnostics.addAll(requestBodyGenerator.getDiagnostics());
                 return null;
             }
-            List<ParameterNode> rBheaderParameters = requestBodyGenerator.getHeaderParameters();
+            headerParameters = requestBodyGenerator.getHeaderSchemas();
             parameterList.add(requestBody.get());
             parameterList.add(comma);
-            if (!rBheaderParameters.isEmpty()) {
-                rBheaderParameters.forEach(header -> {
-                    if (!paramName.contains(header.toString())) {
-                        paramName.add(header.toString());
-                        if (header instanceof DefaultableParameterNode defaultableParameterNode) {
-                            defaultable.add(defaultableParameterNode);
-                            defaultable.add(comma);
-                        } else {
-                            parameterList.add(header);
-                            parameterList.add(comma);
-                        }
-                    }
-                });
+        }
+
+        // 2. parameters - query, requestBody, headers
+        if (parameters != null) {
+            populateHeaderAndQueryParameters(parameters, queryParameters, headerParameters);
+
+            HeadersParameterGenerator headersParameterGenerator = new HeadersParameterGenerator(headerParameters,
+                    openAPI, operation, httpMethod, path);
+            Optional<ParameterNode> headers;
+            if (headerParameters.isEmpty()) {
+                hasDefaultHeader = true;
+                headers = HeadersParameterGenerator.getDefaultParameterNode();
+            } else {
+                headers = headersParameterGenerator.generateParameterNode();
+            }
+
+            if (headers.isPresent()) {
+                hasHeadersParam = true;
+                if (headers.get() instanceof RequiredParameterNode headerNode) {
+                    parameterList.add(headerNode);
+                    parameterList.add(comma);
+                } else {
+                    defaultable.add(headers.get());
+                    defaultable.add(comma);
+                }
+            } else if (!headerParameters.isEmpty()) {
+                diagnostics.addAll(headersParameterGenerator.getDiagnostics());
+                return null;
+            }
+
+            QueriesParameterGenerator queriesParameterGenerator = new QueriesParameterGenerator(queryParameters,
+                    openAPI, operation, httpMethod, path);
+            Optional<ParameterNode> queries = queriesParameterGenerator.generateParameterNode();
+            if (queries.isPresent()) {
+                hasQueriesParam = true;
+                parameterList.add(queries.get());
+                parameterList.add(comma);
+            } else if (!queryParameters.isEmpty()) {
+                diagnostics.addAll(queriesParameterGenerator.getDiagnostics());
+                return null;
+            }
+        } else {
+            ParameterNode defaultHeaderParam = HeadersParameterGenerator.getDefaultParameterNode().orElse(null);
+            if (defaultHeaderParam != null) {
+                hasDefaultHeader = true;
+                defaultable.add(defaultHeaderParam);
+                defaultable.add(comma);
             }
         }
+
         return new ParametersInfo(parameterList, defaultable);
+    }
+
+    private void populateHeaderAndQueryParameters(List<Parameter> parameters, List<Parameter> queryParameters,
+                                                  List<Parameter> headerParameters) {
+        for (Parameter parameter : parameters) {
+            if (parameter.get$ref() != null) {
+                String paramType = null;
+                try {
+                    paramType = extractReferenceType(parameter.get$ref());
+                } catch (BallerinaOpenApiException e) {
+                    ClientDiagnosticImp clientDiagnostic = new ClientDiagnosticImp(OAS_CLIENT_100,
+                            parameter.get$ref());
+                    diagnostics.add(clientDiagnostic);
+                }
+                parameter = openAPI.getComponents().getParameters().get(paramType);
+            }
+
+            String in = parameter.getIn();
+
+            switch (in) {
+                case "query":
+                    queryParameters.add(parameter);
+                    break;
+                case "header":
+                    headerParameters.add(parameter);
+                    break;
+                default:
+                    break;
+            }
+        }
     }
 
     protected FunctionReturnTypeGeneratorImp getFunctionReturnTypeGenerator() {
@@ -195,5 +215,17 @@ public class ResourceFunctionSignatureGenerator implements FunctionSignatureGene
     @Override
     public List<ClientDiagnostic> getDiagnostics() {
         return diagnostics;
+    }
+
+    public boolean hasDefaultHeaders() {
+        return hasDefaultHeader;
+    }
+
+    public boolean hasHeaders() {
+        return hasHeadersParam;
+    }
+
+    public boolean hasQueries() {
+        return hasQueriesParam;
     }
 }
