+++ openapi-core/src/main/java/io/ballerina/openapi/core/generators/client/parameter/ParameterGenerator.java
@@ -27,7 +27,6 @@ import java.util.Optional;
 public interface ParameterGenerator {
     //type handler attribute
     Optional<ParameterNode> generateParameterNode();
-    Optional<ParameterNode> generateParameterNode(boolean treatDefaultableAsRequired);
     List<ClientDiagnostic> getDiagnostics();
 
 }
