+++ openapi-core/src/main/java/io/ballerina/openapi/core/generators/client/parameter/PathParameterGenerator.java
@@ -50,11 +50,6 @@ public class PathParameterGenerator implements ParameterGenerator {
 
     @Override
     public Optional<ParameterNode> generateParameterNode() {
-        return generateParameterNode(false);
-    }
-
-    @Override
-    public Optional<ParameterNode> generateParameterNode(boolean treatDefaultableAsRequired) {
         IdentifierToken paramName = createIdentifierToken(escapeIdentifier(parameter.getName()));
         // type should be a any type node.
         Schema parameterSchema = parameter.getSchema();
